chore: initial setup
I copied over code from my previous entries and updated them to my
latest level of knowledge. That meant REUSE as well as porting the tests
over to Vitest. It builds. The Husky setup changed a little as I am
going to use it for a more strict quality assurance.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>