chore: toot the current bundle size
I ported a script from 2022 that toots the bundle size to my GoToSocial
profile on games whenever I create a new ZIP archive.

This should motivate me to do this step regularly and have a track
record.
Compared to 2022 the masto package changed its API as response to a
tighter behaviour on Mastodon's end. That's cool.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>