feat: first version of prompt
Since the one category is about AI and folks generally expect a text
input for those generative AI nonsense, give them a text input.

Due to constant redrawing I currently loose focus after every input. I
expect this to change once I break the level scene into components.

For now my goal was to implement the learnings I had from videos on
YouTube. In principle, it is breaking the text input on spaces and treat
the first word as a command (verb). Then match the rest.
To improve the user experience I populate a datalist so there should be
suggestions on how to complete the prompt.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>