feat: add a debounce helper
I'm not sure whether I will actually use this but this small function
allows to delay a function call by some time in milliseconds. I was
adding it while investigating a laggy input element (only to find out
that I was barking up the wrong tree eventually). Might be handy, I
don't know. Tree shaking makes sure it is not part of the bundle if I am
not actively using it.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>