test: write tests for refactored code
Now that I am happy with the code I cover it with tests to ensure it
will not break down the road.

I could have used Testing Library for DOM to improve the expect
statements but this approach I took here works, too. Although I like the
getByLabelText() and friends.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>