refactor: replace datalist with combobox
When I tried the game on Firefox for Android I noticed that no
suggestions show up. Searching online I discovered an old bug:

https://bugzilla.mozilla.org/show_bug.cgi?id=1535985

from 2019. Still not addressed. Looking for alternatives I found the
amazing Awesomeplete:

https://projects.verou.me/awesomplete/

which does more than I need (but is accessible). So instead I went with
an HTML4 approach:

https://stackoverflow.com/a/20127724

It works. Maybe not in the most elegant way but it gets the job done.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>