feat: refactor State to use RDF
I'm using JSON-RDF here to associate subjects with predicates and
objects. This is likely not to check out with off-the-shelf RDF tooling
but is easy enough to understand and rewrite for those if someone
fancies. I have the keys in place that the right thing should be
guessable. Unlike proper RDF I struggled with getting the nesting right.
It appears that an object can be complex, as in, not only another URI or
string. I cannot handle those cases.

I'm also disappointed that the URIs I do use (like anchors as keys on
thefacts property) cannot be properly typed as they're "strings". The
union would be discerned on the value of value of rdf:type. Too hard to
teach TypeScript in this particular setup.

Nevertheless I have adjusted the clauses and reducers to work with this
properties now.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>