refactor: use SchemaGameLocation as targets
Now I am using a richer expression to describe target locations that can
be reached.

This one was tricky. Not because of the code but due to the types. You
see, ValuwFlows describes locations as SpatialThing which represents
places on our Earth by longitude, latitude and altitude. That doesn't
work that well with games :b

Instead I looked at schema.org (actually first in the coding session)
and found gameLocation as a Place. That is also pointing to geo
coordinates but accepts text descriptions as well. So by giving Place a
name I can associate nearby maps as Place by name, too. This is more
abstract.

But even then, in a physical description I might also want to describe
the layout of rooms in a building with ValueFlows (or compatible
ontologies) and would prefer a logically form over a physical.

In short: this particular decision is incompatible with ValueFlows. I
don't care much as this is my game and I draw from ontologies as a
shortcut of coming up with different concepts and instead rely on the
wisdowm of the crowd.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>