feat: extend state on meet
Originally this was only introduced another (non-player-)character to
the game world and associated the acquaintance. But then I figured I
might want to hand the new NPC an apple as item that it can gift the
player. He things got complicated.

You see, if I want to apply the ValueFlows ontology things aren't easy.
It is used for Enterprise Resource Planning and similar domains. A
little bit overshooting to what I want here.

In short, things in the game world are described as EconomicResources.
To modify their quantity I have to introduce an EconomicEvent describing
the change. They type of change is represented by an Action.
EconomicEvents involve a provider and a receiver, both are Agents (in vf
as well as foaf vocabuly). A Person is an Agent in FoaF. ValueFlows
introduces EcologicalAgent as another alternative to represent trees,
rivers and such.

This means quite some new properties to be defined.

Once I have a third character in the game I will need to refactor this
reducer by splitting it. One is creating a new character, the others are
associating the items. In order to do that I will need to compute the
name outside of the reducer (in the event handler?) and pass it in as an
Action payload (in the Redux-like sense).

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>