feat: use ValueFlows objects on pickup
This will need a renaming. In ValueFlows a pickup is transporting
EconomicResource from one location to another. As in, loading it on a
PickUp Truck and driving it over.

This is not the pickup I imagined. Taking an apple from the ground or
similar.

I would also need to have a better way to look at the EconomicEvents.
Actually, picking up should create one, I think. So by prompting this a
new event would need to be created.

That makes me think of deciding what EconomicEvent to dispatch on the
event handler level oninput. I could then have pickup, go and other
clauses turned into new Action/Reducer pairs.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>