feat: include quantities
It took me quite some time to grasp how this works. I had the luck that
the units and measured are used in other contexts as well so I was able
to look at examples in the wild. Given my use of JSON-RDF makes the
handling quite awkward. However, I have an implementation now that I can
subsequently refactor to strip away the nitty-gritty details.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>