refactor: extract generateName into helper
This allows me to generate names in different parts of my codebase. For
example, when the level scene is entered I could populate a set of NPCs
now. I would still need to check whether the new name is already in the
facts property of state, but that is the responsibility of the caller.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>