refactor: generate name upfront
This is generating a new name in the event handler. It can potentially
be long (especially in later game) since I always look up whether I have
already used the name.

Another drawback is that I have to pass the seed around. But this is the
price of using a seedable Pseudo Random Number Generator.

On the plus side I don't have to run it within the reducer. But I need
to pass it through so it is tracked in the state for the next
invocation.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>