refactor: generate economic events
The function signature here is bad. Too many arguments. I could wrap
them in an object but those are a little harder to document in JSDoc.

I feel bad to pass so many URIs around. The way I structured it (with
keys ultimately deciding what type of Object the key points to) is
unfortunate. I would love to tell TypeScript by JSDoc what kind of
Object it is. This is the next best thing I could come up with.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>