refactor: extract generate FoaF Person
This is how I envisioned the reducer to look like!

Except it does more than it says. It's not merely tracking the
acquaintance of two people but also creating a new one.

Given that I have factory functions now I can move that part to a new
action/reducer pair and focus on the meeting.

I also loosened the type definitions for foaf:knows here a little. I
allow it to be empty.

Finally I should record that ValueFlows assume foaf:Agent instead of
foaf:Person which is a superclass. I orginally thought of all NPCs to
represent people but given that I have an #AppleTree now I ought to
update the specification. I will need to take another look at foaf
before doing so. I recall that it was an alias.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>