feat: acquaintance people with each other
This is what I always envisioned the reducer to do: add people to each
other's foaf:knows!

It was only a workaround to create the InnHolder upfront.
I should move that out next. I also noticed that the pickup clause (to
be renamed to offer clause) and the offerReducer trigger an meetAction
(to ensure that the State is properly prepared). These are workarounds.
When I have a new action/reducer for adding another person to the game I
can move refactor those action calls, too.

Given that a certain person is always working on some shop, I might add
this occupation I saw in some RDF to tell them apart irrespective of the
random name assigned to them.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>