docs: wrote journal entry
I was unsure how to handle measures first. Sure, there is a numeric part
and an unit. But why do I have data types like NonNegativeInteger in
xsd? It doesn't make sense to me. I found a StackOverflow answer that
wrote it like `value^^unit`. Got to research what that means.

In the meantime I figured out how to put all the required values into
the objects. I noticed that I still had schema:seller here but it was
replaced by vf:provider. Cleaning up the objects by returning them from
factories had a better outcome than I expected. It helped me get more
clarity on the intent of a function.

I'm on a good way! But the time is going to be a challenge. I should
test the features on itch somewhen soon. They had some feature policy
applied to the iFrame if I recall correctly. Would be a bummer if my
code failed to run there.

In fact, someone reported a crash on the Audio code with TOR browser. At
the very least I could wrap it in a try-catch.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>